ANSI
SCCS-vsn: include/arpa/inet.h 5.6

+++ usr/src/include/arpa/inet.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)inet.h	5.5 (Berkeley) %G%
+ *	@(#)inet.h	5.6 (Berkeley) %G%
  */
 
 /* External definitions for functions in inet(3) */
@@ -14,7 +14,7 @@
 __BEGIN_DECLS
 extern unsigned long	 inet_addr __P((const char *));
 extern unsigned long	 inet_lnaof __P((struct in_addr));
-extern struct in_addr	 inet_makeaddr __P((int , int));
+extern struct in_addr	 inet_makeaddr __P((u_long , u_long));
 extern unsigned long	 inet_netof __P((struct in_addr));
 extern unsigned long	 inet_network __P((const char *));
 extern char		*inet_ntoa __P((struct in_addr));
